feat: fix Firecrawl tools returning 'not configured' by loading .env before env var check
Bun's auto .env loading only runs from CWD, so 'process.env.FIRECRAWL_API_KEY'
was undefined when the CLI ran from outside the server directory. Every Firecrawl
tool (search, scrape, map) and the chat tool selector would see 'not configured'
even though the key existed in server/.env.

- Refactored src/lib/load-env.ts into an exported callable 'loadEnvOnce()' that
  walks up to 7 directory candidates to find '.env'
- Added 'loadEnvOnce()' call in firecrawl-search, firecrawl-scrape, firecrawl-map,
  web-search, and chat.ts before reading their respective env vars
- Removed the duplicated inline 'loadEnvOnce()' from web-search.ts (~50 lines)
