+++ apps/supercode-cli/server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "supercode-cli",
-  "version": "0.1.32",
+  "version": "0.1.33",
   "description": "AI-powered coding agent CLI",
   "main": "dist/main.js",
   "bin": {
