+++ apps/supercode-cli/server/src/tools/definitions/firecrawl-search.ts
@@ -1,4 +1,5 @@
 import { z } from "zod"
+import { loadEnvOnce } from "../../lib/load-env"
 
 const FIRECRAWL_BASE = "https://api.firecrawl.dev/v2"
 
@@ -32,6 +33,7 @@ export const firecrawlSearchTool = {
     "If success is false, do NOT invent search results — relay the error to the user.",
   parameters: firecrawlSearchSchema,
   execute: async ({ query, maxResults, includeDomains, excludeDomains }: FirecrawlSearchArgs): Promise<string> => {
+    loadEnvOnce()
     const apiKey = process.env.FIRECRAWL_API_KEY
 
     if (!apiKey) {
